3rd-party-api-rate-limit/requirements.txt
aws-cdk-lib==2.106.1
constructs>=10.0.0,<11.0.0
boto3>=1.33.4
